+++ crates/factorio-api/src/lib.rs
@@ -460,6 +460,7 @@ impl LuaAny {
 include!(concat!(env!("OUT_DIR"), "/mod.rs"));
 
 pub mod lua_libs;
+pub mod prototypes;
 pub mod settings;
 
 pub use event_filters::{EventFilterEntry, FilterMethodSpec, filter_method_spec};
@@ -479,6 +480,7 @@ pub mod prelude {
     pub use crate::lua_fn;
     pub use crate::lua_fn0;
     pub use crate::lua_fn2;
+    pub use crate::prototypes::Item;
     pub use crate::settings::{
         BoolSetting, DoubleSetting, IntSetting, ModSettingValue, StringSetting, data, settings,
     };