+++ crates/factorio-rs/src/lib.rs
@@ -2,8 +2,8 @@
 
 pub use factorio_api::{self, event_type_to_name};
 pub use factorio_macros::{
-    control, data, data_final_fixes, data_updates, event, export, locale, mod_settings, settings,
-    settings_final_fixes, settings_updates, shared,
+    control, data, data_final_fixes, data_updates, event, export, item, locale, mod_settings,
+    settings, settings_final_fixes, settings_updates, shared,
 };
 pub use factorio_macros::{
     control_mod, data_final_fixes_mod, data_mod, data_updates_mod, settings_final_fixes_mod,
@@ -24,9 +24,9 @@ pub mod test;
 pub mod prelude {
     pub use crate::{
         control, control_mod, data, data_final_fixes, data_final_fixes_mod, data_mod, data_updates,
-        data_updates_mod, event, export, locale, mod_settings, settings, settings_final_fixes,
-        settings_final_fixes_mod, settings_mod, settings_updates, settings_updates_mod, shared,
-        shared_mod,
+        data_updates_mod, event, export, item, locale, mod_settings, settings,
+        settings_final_fixes, settings_final_fixes_mod, settings_mod, settings_updates,
+        settings_updates_mod, shared, shared_mod,
     };
     pub use factorio_api::prelude::*;